cart-service/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
requests==2.31.0
django-cors-headers==4.0.0
PyMySQL==1.1.0
cryptography

pika==1.3.2
PyJWT==2.8.0
